+++ page/pager.go
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package gxpage
+
+// Page is the abstraction for pagination usage.
+type Pager interface {
+
+	// OffSet will return the offset
+	Offset() int
+
+	// GetPageSize will return the page size
+	GetPageSize() int
+
+	// GetTotalPages will return the number of total pages
+	GetTotalPages() int
+
+	// GetData will return the data
+	GetData() []interface{}
+
+	// GetDataSize will return the size of data.
+	// Usually it's len(GetData())
+	GetDataSize() int
+
+	// HasNext will return whether has next page
+	HasNext() bool
+
+	// HasData will return whether this page has data.
+	HasData() bool
+}
